Moved duplicated code from the drivers' electrify function to the base electrify function.
