+++ src/drivers/cordova-sqlite-storage/test.ts
@@ -44,12 +44,14 @@ export const initTestable = async <
     dbName,
     dbSchemas,
     adapter,
-    migrator,
-    notifier,
     socketFactory,
-    console,
-    registry,
-    config
+    config,
+    {
+      notifier: notifier,
+      migrator: migrator,
+      console: console,
+      registry: registry,
+    }
   )
 
   return [db, notifier, dal]
