+++ src/drivers/react-native-sqlite-storage/index.ts
@@ -8,12 +8,8 @@ import {
   electrify as baseElectrify,
 } from '../../electric/index'
 
-import { BundleMigrator } from '../../migrators/bundle'
-import { EventNotifier } from '../../notifiers/event'
-import { globalRegistry } from '../../satellite/registry'
-
 import { DatabaseAdapter } from './adapter'
-import { ElectricConfig, hydrateConfig } from '../../config'
+import { ElectricConfig } from '../../config'
 import { WebSocketReactNativeFactory } from '../../sockets/react-native'
 import { Database } from './database'
 import { DalNamespace, DbSchemas } from '../../client/model/dalNamespace'
@@ -34,8 +30,6 @@ import uuid from 'react-native-uuid'
     : global
 )
 
-import { ConsoleHttpClient } from '../../auth'
-
 export { DatabaseAdapter }
 
 export const electrify = async <T extends Database, S extends DbSchemas>(
@@ -46,26 +40,16 @@ export const electrify = async <T extends Database, S extends DbSchemas>(
   opts?: ElectrifyOptions
 ): Promise<DalNamespace<S>> => {
   const dbName: DbName = db.dbName
-  const configWithDefaults = hydrateConfig(config)
-
   const adapter = opts?.adapter || new DatabaseAdapter(db, promisesEnabled)
-  const migrator =
-    opts?.migrator || new BundleMigrator(adapter, config.migrations)
-  const notifier = opts?.notifier || new EventNotifier(dbName)
   const socketFactory = opts?.socketFactory || new WebSocketReactNativeFactory()
-  const console = opts?.console || new ConsoleHttpClient(configWithDefaults)
-  const registry = opts?.registry || globalRegistry
 
   const namespace = await baseElectrify(
     dbName,
     dbSchemas,
     adapter,
-    migrator,
-    notifier,
     socketFactory,
-    console,
-    registry,
-    configWithDefaults
+    config,
+    opts
   )
 
   return namespace
