+++ src/drivers/react-native-sqlite-storage/test.ts
@@ -49,12 +49,14 @@ export const initTestable = async <
     dbName,
     dbSchemas,
     adapter,
-    migrator,
-    notifier,
     socketFactory,
-    console,
-    registry,
-    config
+    config,
+    {
+      notifier: notifier,
+      migrator: migrator,
+      console: console,
+      registry: registry,
+    }
   )
 
   return [db, notifier, dal]
